requirements.txt
claude-agent-sdk>=0.1.48
anthropic>=0.50.0
fastapi>=0.115.0
uvicorn>=0.34.0
websockets>=14.0
httpx>=0.28.0
aiosqlite>=0.21.0
pyyaml>=6.0
python-dotenv>=1.0.0
mcp>=1.0.0
pydantic>=2.10.0
python-multipart>=0.0.9
